+++ src/bludgeon/bludgeon.cpp
@@ -64,6 +64,9 @@ enum WaysToFuckUpTheTag : std::uint64_t {
     /** Fix uppercase references */
     INVALID_UPPERCASE_REFERENCES        = 1ull << 13,
     
+    /** Fix excessive script node count (typically caused by Open Sauce) */
+    EXCESSIVE_SCRIPT_NODES              = 1ull << 13,
+    
     /** Attempt to unfuck anything that can be unfucked (CAUTION: you can unscrew a lightbulb; you can't unscrew a Halo tag) */
     EVERYTHING                          = ~0ull
 };
@@ -82,6 +85,7 @@ enum WaysToFuckUpTheTag : std::uint64_t {
 #define INVALID_SOUND_BUFFER_FIX "incorrect-sound-buffer"
 #define INVALID_INDICES_FIX "invalid-indices"
 #define INVALID_NORMALS_FIX "nonnormal-vectors"
+#define EXCESSIVE_SCRIPT_NODES_FIX "excessive-script-nodes"
 #define BROKEN_STRINGS_FIX "invalid-strings"
 #define EVERYTHING_FIX "everything"
 
@@ -131,6 +135,7 @@ static int bludgeon_tag(const std::filesystem::path &file_path, std::uint64_t fi
             check_fix(broken_normals, "problematic nonnormal vectors detected; fix with " INVALID_NORMALS_FIX);
             check_fix(broken_strings, "problematic strings detected; fix with " BROKEN_STRINGS_FIX);
             check_fix(uppercase_references, "uppercase references detected; fix with " INVALID_UPPERCASE_REFERENCES_FIX);
+            check_fix(excessive_script_nodes, "excessive script nodes detected; fix with " EXCESSIVE_SCRIPT_NODES_FIX);
             
             #undef check_fix
         }
@@ -151,6 +156,7 @@ static int bludgeon_tag(const std::filesystem::path &file_path, std::uint64_t fi
             apply_fix(broken_normals, INVALID_NORMALS, INVALID_NORMALS_FIX);
             apply_fix(broken_strings, BROKEN_STRINGS, BROKEN_STRINGS_FIX);
             apply_fix(uppercase_references, INVALID_UPPERCASE_REFERENCES, INVALID_UPPERCASE_REFERENCES_FIX);
+            apply_fix(excessive_script_nodes, EXCESSIVE_SCRIPT_NODES, EXCESSIVE_SCRIPT_NODES_FIX);
             
             #undef apply_fix
         }
@@ -191,7 +197,7 @@ int main(int argc, char * const *argv) {
     options.emplace_back("fs-path", 'P', 0, "Use a filesystem path for the tag path if specifying a tag.");
     options.emplace_back("all", 'a', 0, "Bludgeon all tags in the tags directory.");
     options.emplace_back("threads", 'j', 1, "Set the number of threads to use for parallel bludgeoning when using --all. Default: CPU thread count");
-    options.emplace_back("type", 'T', 1, "Type of bludgeoning. Can be: " BROKEN_ENUMS_FIX ", " BROKEN_RANGE_FIX ", " BROKEN_STRINGS_FIX ", " BROKEN_REFERENCE_CLASSES_FIX ", " INVALID_MODEL_MARKERS_FIX ", " MISSING_SCRIPTS_FIX ", " INVALID_SOUND_BUFFER_FIX ", " INVALID_VERTICES_FIX ", " INVALID_NORMALS_FIX ", " INVALID_UPPERCASE_REFERENCES_FIX ", " NO_FIXES_FIX ", " EVERYTHING_FIX " (default: " NO_FIXES_FIX ")");
+    options.emplace_back("type", 'T', 1, "Type of bludgeoning. Can be: " BROKEN_ENUMS_FIX ", " BROKEN_RANGE_FIX ", " BROKEN_STRINGS_FIX ", " BROKEN_REFERENCE_CLASSES_FIX ", " INVALID_MODEL_MARKERS_FIX ", " MISSING_SCRIPTS_FIX ", " INVALID_SOUND_BUFFER_FIX ", " INVALID_VERTICES_FIX ", " INVALID_NORMALS_FIX ", " INVALID_UPPERCASE_REFERENCES_FIX ", " EXCESSIVE_SCRIPT_NODES_FIX ", " NO_FIXES_FIX ", " EVERYTHING_FIX " (default: " NO_FIXES_FIX ")");
 
     static constexpr char DESCRIPTION[] = "Convinces tags to work with Invader.";
     static constexpr char USAGE[] = "[options] <-a | tag.class>";
@@ -285,6 +291,9 @@ int main(int argc, char * const *argv) {
                 }
                 else if(std::strcmp(arguments[0], INVALID_UPPERCASE_REFERENCES_FIX) == 0) {
                     bludgeon_options.fixes = bludgeon_options.fixes | WaysToFuckUpTheTag::INVALID_UPPERCASE_REFERENCES;
+                }
+                else if(std::strcmp(arguments[0], EXCESSIVE_SCRIPT_NODES_FIX) == 0) {
+                    bludgeon_options.fixes = bludgeon_options.fixes | WaysToFuckUpTheTag::EXCESSIVE_SCRIPT_NODES;
                 }
                 else if(std::strcmp(arguments[0], EVERYTHING_FIX) == 0) {
                     bludgeon_options.fixes = WaysToFuckUpTheTag::EVERYTHING;